==> 01-clean-filter-listings.ipynb <==
import numpy as np
import pandas as pd
from keys import all_craigslist_listings_path

output_path = 'data/craigslist_listings_cleaned.csv'

## Load the data

# function to convert string to float and handle empty string as NaN
def to_float(string_value):
    string_value = string_value.strip()
    return np.float(string_value) if string_value else np.nan

# load the full combined data set, converting numeric columns to float using to_float
converters = {'neighborhood':str, 
              'title':str, 
              'price':to_float, 
              'bedrooms':to_float, 
              'pid':str, 
              'date':str, 
              'link':str, 
              'sqft':to_float, 
              'sourcepage':str, 
              'longitude':to_float, 
              'latitude':to_float}

%%time
all_listings = pd.read_csv(all_craigslist_listings_path, converters=converters)
all_listings = all_listings.rename(columns={'price':'rent', 'longitude':'lng', 'latitude':'lat'})
print(len(all_listings))

## Filter the data

# de-dupe rows by PID
df = all_listings.drop_duplicates(subset='pid')
len(df)

# retain only rows containing latlng data
mask = (pd.notnull(df['lng'])) & (pd.notnull(df['lat']))
df = df[mask]
len(df)

# define percentiles for valid values
upper_pctl = 0.998
lower_pctl = 0.002

upper_rent = df['rent'].quantile(q=upper_pctl, interpolation='higher')
lower_rent = df['rent'].quantile(q=lower_pctl, interpolation='lower')

upper_sqft = df['sqft'].quantile(q=upper_pctl, interpolation='higher')
lower_sqft = df['sqft'].quantile(q=lower_pctl, interpolation='lower')

print('valid rent range: {:.0f} - {:.0f}'.format(lower_rent, upper_rent))
print('valid sqft range: {:.0f} - {:.0f}'.format(lower_sqft, upper_sqft))

# filter out any rows with rent or sqft outside of the reasonable values (but keep nulls)
rent_mask = ((df['rent'] > lower_rent) & (df['rent'] < upper_rent)) | (pd.isnull(df['rent']))
sqft_mask = ((df['sqft'] > lower_sqft) & (df['sqft'] < upper_sqft)) | (pd.isnull(df['sqft']))
df = pd.DataFrame(df[rent_mask & sqft_mask])
len(df)

# filter out any rows with duplicate (and non-null) values across all of these columns
not_null = pd.notnull(df['rent']) & pd.notnull(df['bedrooms'])
duplicated = df.duplicated(subset=['rent', 'sqft', 'bedrooms', 'neighborhood'], keep='first')
df = df[~(not_null & duplicated)]
len(df)

## Create and clean up features

# extract the subdomain/region from the link column
df['region'] = df['link'].str.extract('http://(.*).craigslist.org', expand=False)

# drop unnecessary columns and reindex
df = df.drop(columns=['title', 'neighborhood', 'link', 'sourcepage', 'pid'])
df = df.reindex(columns=['date', 'region', 'bedrooms', 'rent', 'sqft', 'lat', 'lng'])

df.head()

## Save to disk

%%time
df.to_csv(output_path, index=False, encoding='utf-8')
print(output_path)
