==> 04-select-tracts-by-cities.ipynb <==
# Get all tracts within certain cities

Given a CSV file containing city names, get all the tracts within those cities' boundaries.

import geopandas as gpd
import json
import os
import pandas as pd

all_tracts_path = 'data/us_census_tracts_2014'
places_path = 'data/us_census_places_2014'
states_by_fips_path = 'data/states_by_fips.json'
cities_path = 'data/study_sites.csv'
output_path = 'data/tracts_in_cities_study_area.geojson'

# load the city names that make up our study sites
study_sites = pd.read_csv(cities_path, encoding='utf-8')
len(study_sites)

%%time
# load all US census tracts shapefile
all_tracts = gpd.read_file(all_tracts_path)
len(all_tracts)

%%time
# load all US places (cities/towns) shapefile
places = gpd.GeoDataFrame()
for folder in os.listdir(places_path):
    path = '{}/{}'.format(places_path, folder)
    gdf_tmp = gpd.read_file(path)
    places = places.append(gdf_tmp)
len(places)

# get state abbreviation from FIPS
with open(states_by_fips_path) as f:
    states = json.load(f)
fips_state = {k:v['abbreviation'] for k, v in states.items()}
places['state'] = places['STATEFP'].replace(fips_state, inplace=False)

cities_states = study_sites.apply(lambda row: '{}, {}'.format(row['city'], row['state']), axis=1)

# find these city names in the GDF of all census places
gdf_cities = gpd.GeoDataFrame()
for city_state in cities_states:
    city, state = [item.strip() for item in city_state.split(',')]
    
    mask = (places['NAME']==city) & (places['state']==state)
    if not mask.sum()==1:
        mask = (places['NAME'].str.contains(city)) & (places['state']==state)
        if not mask.sum()==1:
            mask = (places['NAME'].str.contains(city)) & (places['state']==state) & ~(places['NAMELSAD'].str.contains('CDP'))
            if not mask.sum()==1:
                print('Cannot uniquely find "{}"'.format(city_state))
                
    gdf_city = places.loc[mask]
    gdf_cities = gdf_cities.append(gdf_city)
    
len(gdf_cities)

# make "name" field like "city, state"
gdf_cities['name'] = gdf_cities.apply(lambda row: '{}, {}'.format(row['NAME'], row['state']), axis=1)
gdf_cities['name'] = gdf_cities['name'].replace({'Indianapolis city (balance), IN' : 'Indianapolis, IN',
                                                 'Nashville-Davidson metropolitan government (balance), TN' : 'Nashville, TN'})

# make gdf of the cities for joining
cities = gdf_cities[['GEOID', 'name', 'geometry']]
cities = cities.rename(columns={'GEOID':'place_geoid', 'name':'place_name'})
cities = cities.set_index('place_geoid')

# make gdf of the tracts for joining
tract_geoms = all_tracts.set_index('GEOID')[['geometry', 'ALAND']]

%%time
# shrink tracts by ~1 meter to avoid peripheral touches on the outside of the city boundary
tract_geoms['geom_tmp'] = tract_geoms['geometry'].buffer(-0.00001)
tract_geoms = tract_geoms.set_geometry('geom_tmp')

%%time
assert tract_geoms.crs == cities.crs
tracts = gpd.sjoin(tract_geoms, cities, how='inner', op='intersects')
print(len(tracts))

# remove the temporary shrunken geometry
tracts = tracts.set_geometry('geometry').drop(columns=['geom_tmp'])
tracts = tracts.rename(columns={'index_right':'place_geoid'})
tracts.head()

%%time
gdf_save = tracts.reset_index().rename(columns={'index':'GEOID'})
os.remove(output_path) # due to overwriting bug in fiona
gdf_save.to_file(output_path, driver='GeoJSON')
print(output_path)
