==> 05-download-census-data.ipynb <==
# Download census tract data from API

  - Available data: https://api.census.gov/data.html
  - Variables (aka, fields) you can query for: https://api.census.gov/data/2014/acs5/profile/variables.html
  - Browse variables for a place here: https://www.census.gov/acs/www/data/data-tables-and-tools/data-profiles/2014/
  - Sample query: https://api.census.gov/data/2014/acs5/profile?get=DP05_0001E&for=tract:400100&in=state:06+county:001
  - County FIPS codes: https://www.nrcs.usda.gov/wps/portal/nrcs/detail/?cid=nrcs143_013697
  
#### Variables name format

More info: https://www.census.gov/data/developers/data-sets/acs-5year/data-notes.html

variable name format: [TableID]_[RowNumber][VariableType]

Example: Variable DP02_0002PE, "Family households (families)", represents the percent estimate for table DP02 row number 2.

DP (Data Profile): Table type containing broad social, economic, housing, and demographic information in a total of four profiles.

  - DP02: Social Characteristics — includes Education, Marital Status, Relationships, Fertility, Grandparents... 
  - DP03: Economic Characteristics — includes Income, Employment, Occupation, Commuting to Work... 
  - DP04: Housing Characteristics — includes Occupancy and Structure, Housing Value and Costs, Utilities... 
  - DP05: Demographic Characteristics — includes Sex and Age, Race, Hispanic Origin, Housing Units... 

Variable suffixes:

  - E = estimate
  - M = margin of error
  - PE = percent estimate (of total)
  - PM = margin of error for corresponding PE
  - A = annotation

import geopandas as gpd
import getcensus as gc
import os
import pandas as pd
from keys import census_api_key

# which census dataset
dataset = 'acs5'

# which vintage year
year = 2014

# which census variables to retrieve for each tract
variables = ['DP05_0001E',    #total pop
             'DP05_0017E',    #median age
             'DP05_0008PE',   #pct pop aged 20-24
             'DP05_0009PE',   #pct pop aged 25-34
             'DP05_0021PE',   #pct pop age 65+
             'DP05_0066PE',   #pct pop hispanic or latino
             'DP05_0072PE',   #pct pop non-hispanic white alone
             'DP05_0033PE',   #pct pop black
             'DP05_0039PE',   #pct pop asian
             'DP05_0002PE',   #pct pop male
             'DP04_0007PE',   #pct single family detached homes
             'DP04_0088E',    #median value of owner occupied units (dollars)
             'DP04_0036E',    #median number of rooms in house
             'DP04_0025PE',   #pct structure built 1939 or earlier
             'DP04_0046PE',   #pct renter-occupied housing units
             'DP04_0046E',    #count renter-occupied housing units
             'DP04_0005E',    #rental vacancy rate
             'DP04_0048E',    #average household size of renter-occupied housing units
             'DP04_0132E',    #median gross rent (dollars)
             'DP04_0139PE',   #pct with gross rent 30-34.9% of household income
             'DP04_0140PE',   #pct with gross rent 35% or more of household income
             'DP03_0062E',    #median household income
             'DP03_0025E',    #mean travel time to work
             'DP03_0019PE',   #pct commute drove alone
             'DP03_0128PE',   #pct people with income below povery level
             'DP02_0057PE',   #pct who are students currently enrolled in college or grad school
             'DP02_0022PE',   #pct population in household is nonrelatives
             'DP02_0079PE',   #pct residence 1 year ago was same house
             'DP02_0067PE',   #pct bachelor's degree or higher
             'DP02_0111PE',   #pct with english only language spoken at home
             'DP02_0092PE']   #pct of population foreign born

# data directories
tracts_path = 'data/tracts_in_cities_study_area.geojson'
output_path = 'data/downloaded_census_data.geojson'

# download and display census descriptions of each variable
variable_descriptions = gc.get_census_variable_descriptions(dataset=dataset, 
                                                            year=year, 
                                                            variables=variables)
for v, d in variable_descriptions.items():
    print('{}\t{}'.format(v, d['label']))

## Get vars from ACS DP

# load the tracts in our study areas
gdf = gpd.read_file(tracts_path).sort_values(by='GEOID')
len(gdf)

%%time
df = gc.get_census_tracts_data(tract_fips=gdf['GEOID'], api_key=census_api_key, dataset=dataset,
                               year=year, variables=variables, clean=True)

%%time
# get 2012 pct white so we can examine its change over time
dataset = 'acs5'
year = 2012
variables = ['DP05_0072PE'] #pct pop non-hispanic white alone
df_2012 = gc.get_census_tracts_data(tract_fips=gdf['GEOID'], api_key=census_api_key, dataset=dataset,
                                    year=year, variables=variables, clean=True)

%%time
# get 2015 pct white so we can examine its change over time
dataset = 'acs5'
year = 2015
variables = ['DP05_0072PE'] #pct pop non-hispanic white alone
df_2015 = gc.get_census_tracts_data(tract_fips=gdf['GEOID'], api_key=census_api_key, dataset=dataset,
                                    year=year, variables=variables, clean=True)

# merge the 2012 and 2015 variables together into a single df
df_2012 = df_2012.drop(columns=['state', 'county'])
df_2015 = df_2015.drop(columns=['state', 'county'])
df_2012_2015 = pd.merge(left=df_2012, right=df_2015, how='inner', left_index=True, right_index=True,
                        suffixes=('_2012', '_2015'))

# merge the tracts with the 2014 variables, then with the 2012/2015 variables
merged = pd.merge(left=gdf.set_index('GEOID'), right=df, how='inner', left_index=True, right_index=True)
merged = pd.merge(left=merged, right=df_2012_2015, how='inner', left_index=True, right_index=True)
merged.head()

# make sure everything we merged is the same length
assert len(gdf) == len(df) == len(df_2012_2015) == len(merged)

%%time
os.remove(output_path) # due to overwriting bug in fiona
merged.reset_index().to_file(output_path, driver='GeoJSON')
print(output_path)
